fix: rename code executor
